kraken tests and begin oodle refactor
