+++ WolvenKit.CLI/WolvenKit.CLI.csproj
@@ -7,22 +7,22 @@
     <AssemblyVersion>1.5.2</AssemblyVersion>
     <InformationalVersion>1.5.2</InformationalVersion>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-      <PackAsTool>true</PackAsTool>
-      <ToolCommandName>cp77tools</ToolCommandName>
-      <PackageOutputPath>./nupkg</PackageOutputPath>
-      <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
-      <PackageId>WolvenKit.CLI</PackageId>
-      <Company>WolvenKit</Company>
-      <Product>WolvenKit.CLI</Product>
-      <Description>Command line modding tool for Cyberpunk 2077.</Description>
-      <PackageLicenseExpression>MIT</PackageLicenseExpression>
-      <PackageProjectUrl>https://github.com/WolvenKit/Wolven-kit</PackageProjectUrl>
-      <RepositoryUrl>https://github.com/WolvenKit/Wolven-kit</RepositoryUrl>
-      <RepositoryType>git</RepositoryType>
-      <PackageTags>wolvenkit, cyberpunk2077</PackageTags>
-      <StartupObject>WolvenKit.CLI.Program</StartupObject>
-      <SatelliteResourceLanguages>en</SatelliteResourceLanguages>
-      <DebugType>embedded</DebugType>
+    <PackAsTool>true</PackAsTool>
+    <ToolCommandName>cp77tools</ToolCommandName>
+    <PackageOutputPath>./nupkg</PackageOutputPath>
+    <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
+    <PackageId>WolvenKit.CLI</PackageId>
+    <Company>WolvenKit</Company>
+    <Product>WolvenKit.CLI</Product>
+    <Description>Command line modding tool for Cyberpunk 2077.</Description>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <PackageProjectUrl>https://github.com/WolvenKit/Wolven-kit</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/WolvenKit/Wolven-kit</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageTags>wolvenkit, cyberpunk2077</PackageTags>
+    <StartupObject>WolvenKit.CLI.Program</StartupObject>
+    <SatelliteResourceLanguages>en</SatelliteResourceLanguages>
+    <DebugType>embedded</DebugType>
     <PublishSingleFile>false</PublishSingleFile>
     <RuntimeIdentifier>win-x64</RuntimeIdentifier>
     <SelfContained>true</SelfContained>
