--- WolvenKit.Common/Extensions/Cr2wExtensions.cs
@@ -1,15 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using WolvenKit.RED4.Archive.CR2W;
-using WolvenKit.RED4.CR2W;
-
-namespace WolvenKit.Common.Extensions
-{
-    public static class Cr2wExtensions
-    {
-        //public static int GetBufferIndex(this CR2WFile cr2wFile, ICR2WBuffer buffer) => cr2wFile.Buffers.IndexOf(buffer);
-    }
-}
