+++ WolvenKit.Common/RED4/Red4WriterService.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using CP77Tools.Model;
 using WolvenKit.Core.CRC;
+using WolvenKit.Core.Extensions;
 using WolvenKit.Interfaces.Extensions;
 using WolvenKit.RED4.CR2W.Archive;
 using Index = CP77Tools.Model.Index;
