+++ WolvenKit.Common/Tools/DDS/BlockCompression.cs
@@ -1,7 +1,7 @@
 using System.IO;
 using System.Runtime.InteropServices;
 using WolvenKit.Common;
-using WolvenKit.Common.Extensions;
+using WolvenKit.Core.Extensions;
 using WolvenKit.RED4.Archive;
 
 namespace CP77.Common.Image
