--- WolvenKit.Common/Tools/Oodle/DecompressionException.cs
@@ -1,15 +1,0 @@
-using System;
-
-namespace WolvenKit.Common.Oodle
-{
-    public class DecompressionException : Exception
-    {
-        #region Constructors
-
-        public DecompressionException(string message) : base(message)
-        {
-        }
-
-        #endregion Constructors
-    }
-}
