+++ WolvenKit.Console.sln
@@ -11,10 +11,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{B904A345
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WolvenKit.CLI", "WolvenKit.CLI\WolvenKit.CLI.csproj", "{A78BEC35-33F4-4E4E-BC0F-D98FE381CFB6}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WolvenKit.CLI", "WolvenKit.CLI", "{E0ED641B-A0A2-4BB7-B266-DE5AD06D6AAB}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WolvenKit", "WolvenKit", "{BE958494-0BB3-4401-93DB-65458FED482B}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{825169A9-0548-4A0A-A1F1-B99BE7D0C23F}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Networking", "Networking", "{2304D155-E6AB-4AE0-969B-C6D0CDC0CD6D}"
@@ -131,13 +127,12 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{A84C71C3-E7BD-4E88-847B-B743E549429A} = {6BC8815E-BFF9-46A5-9535-0E89A765D75A}
-		{A78BEC35-33F4-4E4E-BC0F-D98FE381CFB6} = {E0ED641B-A0A2-4BB7-B266-DE5AD06D6AAB}
 		{825169A9-0548-4A0A-A1F1-B99BE7D0C23F} = {6BC8815E-BFF9-46A5-9535-0E89A765D75A}
 		{2304D155-E6AB-4AE0-969B-C6D0CDC0CD6D} = {6BC8815E-BFF9-46A5-9535-0E89A765D75A}
 		{7199D583-CA10-4035-9014-65158344B33D} = {6BC8815E-BFF9-46A5-9535-0E89A765D75A}
 		{43800A68-B25D-4D52-9FC4-9E3F247A8020} = {825169A9-0548-4A0A-A1F1-B99BE7D0C23F}
 		{EE7C13FD-10C5-4527-A821-B4FB0C6EB3E8} = {2304D155-E6AB-4AE0-969B-C6D0CDC0CD6D}
-		{D019CB01-9948-418A-8C1C-DB66D38BFCDF} = {BE958494-0BB3-4401-93DB-65458FED482B}
+		{D019CB01-9948-418A-8C1C-DB66D38BFCDF} = {6BC8815E-BFF9-46A5-9535-0E89A765D75A}
 		{CEF1CF72-B8CF-4CBB-AA11-473438EDE3F3} = {825169A9-0548-4A0A-A1F1-B99BE7D0C23F}
 		{96B50D87-B772-4E1F-BF6C-44AFDC2CA748} = {B904A345-E6CF-4D6A-9F10-C04F02340B63}
 		{30FF3602-DA2B-4289-A9C6-7D1E1AA35A81} = {A84C71C3-E7BD-4E88-847B-B743E549429A}
