+++ WolvenKit.Core/Compression/DecompressionException.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace WolvenKit.Core.Compression;
+
+public class DecompressionException : Exception
+{
+    #region Constructors
+
+    public DecompressionException(string message) : base(message)
+    {
+    }
+
+    #endregion Constructors
+}
