+++ WolvenKit.Core/Compression/OodleLib.cs
@@ -1,29 +1,11 @@
 using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
-using WolvenKit.Common.Oodle;
 
-namespace WolvenKit.Common.Tools.Oodle
+namespace WolvenKit.Core.Compression
 {
-    //https://stackoverflow.com/a/23623244
-    internal static class NativeMethods
-    {
-        [DllImport("kernel32.dll")]
-        public static extern IntPtr LoadLibrary(string dllToLoad);
-
-        [DllImport("kernel32.dll")]
-        public static extern IntPtr GetProcAddress(IntPtr hModule, string procedureName);
-
-        [DllImport("kernel32.dll")]
-        public static extern bool FreeLibrary(IntPtr hModule);
-    }
 
 
-    public static class OodleLoadLib
+    public static class OodleLib
     {
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
         private delegate int DecompressDelegate(IntPtr buffer, long bufferSize, IntPtr outputBuffer, long outputBufferSize,
@@ -140,6 +122,6 @@ namespace WolvenKit.Common.Tools.Oodle
                 0
             );
 
-        
+
     }
 }
