+++ WolvenKit.MSTests/HashTests.cs
@@ -10,10 +10,11 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using WolvenKit.Common.Extensions;
 using WolvenKit.Common.FNV1A;
 using WolvenKit.Common.Model;
-using WolvenKit.Common.Oodle;
+using WolvenKit.Core.Compression;
 using WolvenKit.Core;
 using WolvenKit.Interfaces.Extensions;
 using WolvenKit.RED4.Archive;
+using WolvenKit.Core.Extensions;
 
 namespace WolvenKit.MSTests
 {
@@ -44,7 +45,7 @@ namespace WolvenKit.MSTests
             
                 // read KARK header
                 var oodleCompression = stream.ReadStruct<uint>();
-                if (oodleCompression != OodleHelper.KARK)
+                if (oodleCompression != Oodle.KARK)
                 {
                     throw new DecompressionException($"Incorrect hash file.");
                 }
@@ -54,7 +55,7 @@ namespace WolvenKit.MSTests
                 // read the rest of the stream
                 var outputbuffer = new byte[outputsize];
                 var inbuffer = stream.ToByteArray(true);
-                OozNative.Kraken_Decompress(inbuffer, inbuffer.Length, outputbuffer, outputbuffer.Length);
+                KrakenNative.Decompress(inbuffer, outputbuffer);
             
 
                 using (var ms = new MemoryStream(outputbuffer))
@@ -127,7 +128,7 @@ namespace WolvenKit.MSTests
 
                 // read KARK header
                 var oodleCompression = stream.ReadStruct<uint>();
-                if (oodleCompression != OodleHelper.KARK)
+                if (oodleCompression != Oodle.KARK)
                 {
                     throw new DecompressionException($"Incorrect hash file.");
                 }
@@ -137,7 +138,7 @@ namespace WolvenKit.MSTests
                 // read the rest of the stream
                 var outputbuffer = new byte[outputsize];
                 var inbuffer = stream.ToByteArray(true);
-                OozNative.Kraken_Decompress(inbuffer, inbuffer.Length, outputbuffer, outputbuffer.Length);
+                KrakenNative.Decompress(inbuffer, outputbuffer);
 
 
                 using (var ms = new MemoryStream(outputbuffer))
@@ -213,7 +214,7 @@ namespace WolvenKit.MSTests
 
                 // read KARK header
                 var oodleCompression = stream.ReadStruct<uint>();
-                if (oodleCompression != OodleHelper.KARK)
+                if (oodleCompression != Oodle.KARK)
                 {
                     throw new DecompressionException($"Incorrect hash file.");
                 }
@@ -223,7 +224,7 @@ namespace WolvenKit.MSTests
                 // read the rest of the stream
                 var outputbuffer = new byte[outputsize];
                 var inbuffer = stream.ToByteArray(true);
-                OozNative.Kraken_Decompress(inbuffer, inbuffer.Length, outputbuffer, outputbuffer.Length);
+                KrakenNative.Decompress(inbuffer, outputbuffer);
 
 
                 using (var ms = new MemoryStream(outputbuffer))
