+++ WolvenKit.MSTests/ModkitConvertTests.cs
@@ -16,8 +16,8 @@ using Catel.IoC;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
 using WolvenKit.Common.Conversion;
-using WolvenKit.Common.Tools;
-using WolvenKit.Interfaces.Extensions;
+using WolvenKit.Core.Compression;
+using WolvenKit.Core.Extensions;
 using WolvenKit.Modkit.RED4;
 using WolvenKit.MSTests.Model;
 using WolvenKit.RED4.Archive.CR2W;
