+++ WolvenKit.Modkit/RED4/Uncook.cs
@@ -10,7 +10,7 @@ using WolvenKit.Common;
 using WolvenKit.Common.DDS;
 using WolvenKit.Common.Extensions;
 using WolvenKit.Common.Model.Arguments;
-using WolvenKit.Common.Oodle;
+using WolvenKit.Core.Compression;
 using WolvenKit.Common.Services;
 using WolvenKit.Modkit.Extensions;
 using WolvenKit.Modkit.RED4.Opus;
